Add arena ID to the spec
We have multiple arenas this year, so we need some way of storing
which arena the scorecard data relates to.  This patch adds a new
required field called "arena_id".  This can be a string or a number,
allowing us maximum flexibility on how we name the arenas in future
years.

The intention is that a match would be represented by the tuple
(arena_id, match_number).