Create the ability to process multiple files
With Python and `libproton`, the startup time of a proton process is significant. It takes too long to be practical to process a full competition's worth of matches with one process per match. As a result, `srcomp` is currently coupled to the internal API of libproton rather than invoking a proton process.

Being able to process multiple files in one run of the program means that all files can be processed while suffering the startup time only once.